requirements.txt
pydantic>=2.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.1
langchain>=0.2.11
langchain-community>=0.2.10
langgraph>=0.2.32
sentence-transformers>=3.0.1
faiss-cpu>=1.8.0
python-docx>=1.1.2
google-generativeai>=0.7.2
uvicorn>=0.30.3
numpy>=1.26.4
rich>=13.7.1
orjson>=3.10.6
fastapi>=0.112.0
requests>=2.32.3
beautifulsoup4>=4.12.3
lxml>=5.2.2
pypdf>=4.2.0
streamlit>=1.36.0
